Support downloading entitlement certificates from candlepin
When a client request comes in, we read the x509 entitlement certificate
to find the order number, which maps to a subscription id in candlepin.
We can then speak to candlepin, and ask it for the upstream entitlement
certificate that backs that subscription id. After getting the cert/key
pair from candlepin, we build a new ssl context using that x509 data to
speak to the cdn.