Rework try/catch block on config file reading
Previously, we were checking if /etc/thumbslug/thumbslug.conf existed, and then
attempted to read. This would cause thumbslug to incorrectly exit if the conf
file existed but was not readable.

This patch allows thumbslug to continue if the config is not readable for any
reason, as long as the default configs shipped in the jar were readable.