Initial proxy support for thumbslug
This patch allows TS to use a proxy server that sits in between it and the CDN.
One example of this would be a proxy server on a corporate network. The
majority of this was written by jbowes, except for the below known bug :)

During the call to candlepin to fetch the upstream cert, if candlepin returns a
200 with bogus cert data, a 502 will correctly be sent to the client, but the
*next* call will fail due to a connection that did not get cleaned up. This
only occurs in proxy mode.