837386: do not package jna with thumbslug
Previously, thumbslug was being packaged with dependency jars. This did not
work for jna, since it relies on native extensions that are not compiled with
the jar.

Instead, use the jar that is supplied with the system. I also performed the
same for netty and commons-codec; I left akuma and oauth as-is for now since
system-level jars are not available on all distributions yet.