837386: do not package jars with thumbslug
Previously, thumbslug was being packaged with dependency jars. This did not
work for jna, since it relies on native extensions that are not compiled with
the jar.

Instead, use the jar that is supplied with the system. I also performed the
same for other dependency jars as well, and set the classpath in the init
script.