Go back to a megajar for use with the buildfile
When we build with ant, we're building to be packaged in an rpm,
so we must follow proper conventions for packaging, including not
bundling deps.

With buildr, we're building to develop and test, so we don't really care
about bundling or not. Go back to bundling, so it's easy to run and test
thumbslug.