Remove dead and unneeded code
In moving to support certv3, thumbslug stopped parsing the entitlement
cert and instead uses a new candlepin api call that provides the
upstream cert given an entitlement UUID. This removes the need for
including our own DER decoder. It also removes the need to asking
candlepin if an entitlement exists before getting the upstream cert, as
the upstream cert request will first check if the entitlement exists.