916895: Fix use of V3 certificates to the CDN.
If a manifest contains a V3 certificate, logic for parsing the key and
cert was very wrong, expecting them to be the only two sections in the
file. V3 certs however also carry entitlement data and a signature.

Instead, identify explicit start to the cert and key sections and pull
them out specifically. Throw the same exception as before if we're
unable to identify any of the begin/end markers.