Fix thumbslug's spec tests.
The entitlement certificates that the tests used were expired, so
I generated new ones that will last until 2030 and placed all the
necessary information for the candlepin CA in the CA directory.

Additionally, the tests were not running correctly with the -jar
option, so we now just invoke the Main class directly and I altered
the way that flags were passed in to the JVM to make it a little
easier to add things in the future.